fixing errors

+++ frontend/src/Component/WorkoutDetails.jsx
@@ -1,15 +1,25 @@
 import React from 'react'
 import { useWorkoutContext } from '../hooks/UseWorkoutContent'
+import { useAuthContext } from '../hooks/UseAuthContext'
 
 //date-fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({workout}) => {
     const { dispatch } = useWorkoutContext()
+    const {user} = useAuthContext()
+
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
+
         const response = await fetch('http://localhost:5010/api/Workout/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}` 
+            }
         })
         const json = await response.json()
         // console.log(json);
